separate behavior scripts
